requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
shapely>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
